backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
pandas>=1.5.0
numpy>=1.21.0
xgboost>=1.7.0
scikit-learn>=1.1.0
google-cloud-aiplatform>=1.36.0
google-cloud-bigquery>=3.12.0
google-cloud-storage>=2.10.0
google-generativeai>=0.3.0
google-auth>=2.23.0
websockets>=12.0
asyncio>=3.4.3
gunicorn>=21.2.0
optuna>=3.4.0
psutil>=5.9.0
